% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_R_functions.R
\name{import_all}
\alias{import_all}
\title{Import All Function}
\usage{
import_all(df = FALSE)
}
\arguments{
\item{df}{Do you want individual dataframes? Defaults to FALSE}
}
\description{
This function allows you to import all csv and xlsx files in the parent directory.
}
\examples{
import_all()
}
\keyword{csv}
