% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{Log sum of exponentials}
\usage{
log_summed_exps(x = NULL)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
Log of the summation of the exponentials of the numeric values stored in x
}
\description{
Log sum of exponentials
}
\examples{

log_summed_exps(x = 1:2000)
}
