% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_R_functions.R
\name{xlsx_only}
\alias{xlsx_only}
\title{XLSX Import Function}
\usage{
xlsx_only(df = FALSE)
}
\arguments{
\item{df}{Do you want individual dataframes? Defaults to FALSE}
}
\description{
This function allows you to import only xlsx files in the parent directory.
}
\examples{
xlsx_only()
}
\keyword{csv}
